CodyBot-Nonebot/src/plugins/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-cody"
version = "1.0.11"
description = ""
authors = ["chrisyy <1017975501@qq.com>", "i2cy <i2cy@outlook.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_cody"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc2"
nonebot-adapter-onebot = "^2.1.5"

openai = "^0.25.0"
transformers = "^4.25.1"
pyyaml = "^6.0"
nonebot-plugin-htmlrender="^0.2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
